Move scheduling plan into core
